# tests/fixtures/openapi.snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "M+ Trust Factor API",
    "version": "0.1.0",
    "description": "Public character/comparison routes with SWR refresh, Battle.net OAuth IAM, and permission-protected admin routes (shared admin key retained as emergency fallback)."
  },
...
  "paths": {
...
    "/api/v1/admin/ability-catalog": {
      "get": {
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "query",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "classSlug",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "specSlug",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "DPS",
                "TANK",
                "HEALER"
              ]
            },
            "in": "query",
            "name": "role",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "category",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "ownership",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "availability",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "version",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "validationState",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/score-models": {
      "get": {
        "tags": [
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "models": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "key": {
                            "type": "string"
                          },
                          "version": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "config": {
                            "type": "object",
                            "additionalProperties": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "activatedAt": {
                            "type": [
                              "null",
                              "string"
                            ]
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "minLength": 1
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "key": {
                        "type": "string"
                      },
                      "version": {
                        "type": "integer",
                        "minimum": 1
                      },
                      "weights": {
                        "type": "object",
                        "properties": {
                          "performance": {
                            "type": "number"
                          },
                          "survival": {
                            "type": "number"
                          },
                          "utility": {
                            "type": "number"
                          },
                          "experienceConsistency": {
                            "type": "number"
                          },
                          "mythicRaid": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "performance",
                          "survival",
                          "utility",
                          "experienceConsistency",
                          "mythicRaid"
                        ]
                      },
                      "authenticityBlend": {
                        "type": "object",
                        "properties": {
                          "skillWeight": {
                            "type": "number"
                          },
                          "authenticityWeight": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "skillWeight",
                          "authenticityWeight"
                        ]
                      },
                      "confidenceNeutralScore": {
                        "type": "number"
                      },
                      "gradeThresholds": {
                        "type": "object",
                        "properties": {
                          "S": {
                            "type": "number"
                          },
                          "A": {
                            "type": "number"
                          },
                          "B": {
                            "type": "number"
                          },
                          "C": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "S",
                          "A",
                          "B",
                          "C"
                        ]
                      }
                    },
                    "required": [
                      "key",
                      "version",
                      "weights",
                      "authenticityBlend",
                      "confidenceNeutralScore",
                      "gradeThresholds"
                    ]
                  }
                },
                "required": [
                  "key",
                  "name",
                  "config"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "version": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "config": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "activatedAt": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/score-models/{id}/clone": {
      "post": {
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "version": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "config": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "activatedAt": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/score-models/{id}/validate": {
      "post": {
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/score-models/{id}/backtest": {
      "post": {
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scoreModelId": {
                      "type": "string"
                    },
                    "sampleSize": {
                      "type": "number"
                    },
                    "gradeDistribution": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "meanScore": {
                      "type": "number"
                    },
                    "generatedAt": {
                      "type": "string"
                    },
                    "note": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}